Separate upcoming events provider for widget
