+++ android_mobile/src/main/java/com/alexstyl/specialdates/images/ImageLoader.java
@@ -22,7 +22,7 @@ public interface ImageLoader {
     }
 
     interface FixedSizeRequest {
-        Optional<Bitmap> async();
+        Optional<Bitmap> synchronously();
 
         void into(ImageLoadedConsumer consumer);
     }
