+++ android_mobile/src/main/java/com/alexstyl/specialdates/upcoming/widget/list/CircularAvatarFactory.java
@@ -34,7 +34,7 @@ final class CircularAvatarFactory {
         Optional<Bitmap> bitmapOptional = imageLoader
                 .load(contact.getImagePath())
                 .withSize(targetSize, targetSize)
-                .async();
+                .synchronously();
         if (!bitmapOptional.isPresent()) {
             return Optional.absent();
         }
